requirements.txt
flask
python-docx
PyPDF2
pytest